# Add new `Performance/AncestorsInclude` cop