--- a/docs/modules/ROOT/pages/cops.adoc
+++ b/docs/modules/ROOT/pages/cops.adoc
@@ -2,6 +2,7 @@
 
 = Department xref:cops_performance.adoc[Performance]
 
+* xref:cops_performance.adoc#performanceancestorsinclude[Performance/AncestorsInclude]
 * xref:cops_performance.adoc#performancebindcall[Performance/BindCall]
 * xref:cops_performance.adoc#performancecaller[Performance/Caller]
 * xref:cops_performance.adoc#performancecasewhensplat[Performance/CaseWhenSplat]