Related link - https://github.com/JuanitoFatas/fast-ruby#ancestorsinclude-vs--code

### Benchmark

```ruby
# frozen_string_literal: true

require 'bundler/inline'

gemfile(true) do
  gem 'benchmark-ips'
  gem 'benchmark-memory'
end

def fast
  (Class <= Class)
  (Class <= Module)
  (Class <= Object)
  (Class <= Kernel)
  (Class <= BasicObject)
end

def slow
  Class.ancestors.include?(Class)
  Class.ancestors.include?(Module)
  Class.ancestors.include?(Object)
  Class.ancestors.include?(Kernel)
  Class.ancestors.include?(BasicObject)
end

puts('********* IPS *********')

Benchmark.ips do |x|
  x.report('less than or equal') { fast }
  x.report('ancestors.include?') { slow }
  x.compare!
end

puts "********* MEMORY *********"

Benchmark.memory do |x|
  x.report('less than or equal') { fast }
  x.report('ancestors.include?') { slow }
  x.compare!
end
```

### Results

```
********* IPS *********
Warming up --------------------------------------
  less than or equal   444.389k i/100ms
  ancestors.include?    67.737k i/100ms
Calculating -------------------------------------
  less than or equal      4.947M (± 5.9%) i/s -     24.886M in   5.049943s
  ancestors.include?    641.527k (±27.5%) i/s -      2.709M in   5.082782s

Comparison:
  less than or equal:  4947152.6 i/s
  ancestors.include?:   641526.9 i/s - 7.71x  (± 0.00) slower

********* MEMORY *********
Calculating -------------------------------------
  less than or equal     0.000  memsize (     0.000  retained)
                         0.000  objects (     0.000  retained)
                         0.000  strings (     0.000  retained)
  ancestors.include?     1.000k memsize (     0.000  retained)
                         5.000  objects (     0.000  retained)
                         0.000  strings (     0.000  retained)

Comparison:
  less than or equal:          0 allocated
  ancestors.include?:       1000 allocated - Infx more
```